Documentation and refactoring for code related to cinema hall

+++ src/main/java/net/cinemaApplication/backend/service/implementations/CinemaHallImplementation.java
@@ -2,11 +2,12 @@ package net.cinemaApplication.backend.service.implementations;
 
 import net.cinemaApplication.backend.entity.cinemaHall.CinemaHall;
 import net.cinemaApplication.backend.entity.cinemaHall.Seat;
-import net.cinemaApplication.backend.entity.movieSession.MovieSession;
+import net.cinemaApplication.backend.entity.movie.Movie;
 import net.cinemaApplication.backend.repository.CinemaHallRepository;
 import net.cinemaApplication.backend.repository.MovieSessionRepository;
 import net.cinemaApplication.backend.repository.SeatRepository;
 import net.cinemaApplication.backend.service.services.CinemaHallService;
+import net.cinemaApplication.backend.service.services.MovieService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -14,6 +15,8 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 @Service
 public class CinemaHallImplementation implements CinemaHallService {
     @Autowired
@@ -23,9 +26,38 @@ public class CinemaHallImplementation implements CinemaHallService {
     MovieSessionRepository movieSessionRepository;
     @Autowired
     SeatRepository seatRepository;
+
+    /**
+     * @return List of all the cinema halls created
+     */
+    @Override
+    public List<CinemaHall> getAllCinemaHalls() {
+        return cinemaHallRepository.findAll();
+    }
+
+    /**
+     *
+     * @param id of the cinema hall
+     * @return an optional of the cinema hall
+     */
+    @Override
+    public Optional<CinemaHall> getCinemaHallById(Long id) {
+        return cinemaHallRepository.findById(id);
+    }
+
+
+    /**
+     * Method for adding new cinema hall to the system
+     * While creating a new hall, the system automatically adds seats according to cinema hall rows and columns
+     * @param cinemaHall that client wants to create
+     * @return created cinema hall
+     */
     @Override
     public CinemaHall addCinemaHall(CinemaHall cinemaHall) {
-        if (cinemaHallRepository.findAll().stream().anyMatch(c -> c.getHallNr() == cinemaHall.getHallNr())) {
+        if(cinemaHall.getHallNr() < 0) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Hall number must be positive");
+        }
+        if (cinemaHallRepository.findByHallNr(cinemaHall.getHallNr()).isPresent()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Hall nr is already taken");
         }
         if (cinemaHall.getSeatColumns() < 0) {
@@ -36,56 +68,32 @@ public class CinemaHallImplementation implements CinemaHallService {
         }
         CinemaHall savedCinemaHall = cinemaHallRepository.save(cinemaHall);
 
-        addSeatsToTheMovieHall(savedCinemaHall);
+        addSeatsToTheMovieHall(savedCinemaHall); //adds seats to the cinema hall
         return savedCinemaHall;
     }
 
 
-    @Override
-    public List<CinemaHall> getAllCinemaHalls() {
-        return cinemaHallRepository.findAll();
-    }
-
-    @Override
-    public CinemaHall updateCinemaHall(CinemaHall cinemaHall, Long id) {
-        CinemaHall cinemaHallFromDb = cinemaHallRepository.findById(id).get(); //get old cinema hall from database
-
-        //hall nr has to be unique
-        if (cinemaHall.getHallNr() != cinemaHallFromDb.getHallNr()) {
-            if (cinemaHallRepository.findAll().stream().anyMatch(c -> c.getHallNr() == cinemaHall.getHallNr())) {
-                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Hall nr is already taken");
-            }
-        }
-        if (cinemaHall.getSeatColumns() > 0) {
-            cinemaHallFromDb.setSeatColumns(cinemaHall.getSeatColumns());
-        }
-        if (cinemaHall.getSeatRows() > 0) {
-            cinemaHallFromDb.setSeatRows(cinemaHall.getSeatRows());
-        }
-        return cinemaHallRepository.save(cinemaHallFromDb);
-    }
-
+    /**
+     * deletes cinema hall from the system (also deletes movie sessions and tickets related to this)
+     * @param id represents cinema hall id
+     */
     @Override
     public void deleteById(Long id) {
+        if (cinemaHallRepository.findById(id).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cinema hall not found");
+        }
         cinemaHallRepository.deleteById(id);
     }
 
-//    @Override //add hall for the movie session, movie session has to exist already
-//    public MovieSession addHallForTheMovieSession(Long cinemaHallId, Long movieSessionId) {
-//        MovieSession movieSession = movieSessionRepository.findById(movieSessionId)
-//                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Movie session not found"));
-//        CinemaHall cinemaHall = cinemaHallRepository.findById(cinemaHallId)
-//                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Cinema hall not found"));
-//        movieSession.setHall(cinemaHall);
-//        cinemaHall.getSessions().add(movieSession);
-//        cinemaHallRepository.save(cinemaHall);
-//        return movieSession;
-//    }
+    /**
+     * Method for adding new seats to the cinema hall. Method is used while creating a new hall
+     * @param cinemaHall to add seats to
+     */
+    private void addSeatsToTheMovieHall(CinemaHall cinemaHall) {
+        if (cinemaHallRepository.findById(cinemaHall.getId()).isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Cinema hall not found");
+        }
 
-    @Override //add seats to the hall (adds all the seats at one time automatically while creating a hall)
-    public List<Seat> addSeatsToTheMovieHall(CinemaHall cinemaHall) {
-//        CinemaHall cinemaHall = cinemaHallRepository.findById(cinemaHallId)
-//                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Cinema hall not found"));
         int totalSeatsAllowed = cinemaHall.getSeatColumns() * cinemaHall.getSeatRows();
         int existingSeats = cinemaHall.getSeats() != null ? cinemaHall.getSeats().size() : 0;
 
@@ -94,19 +102,15 @@ public class CinemaHallImplementation implements CinemaHallService {
         }
 
         int seatsToAdd = totalSeatsAllowed - existingSeats;
-        List<Seat> newSeats = new ArrayList<>();
         for (int i = 0; i < seatsToAdd; i++) {
             Seat seat = Seat.builder()
                     .seatNr(existingSeats + i + 1)
-//                    .isSeatTaken(false)
                     .hall(cinemaHall)
                     .build();
             cinemaHall.getSeats().add(seat);
-            Seat savedSeat = seatRepository.save(seat);
-            newSeats.add(savedSeat);
+            seatRepository.save(seat);
         }
         cinemaHallRepository.save(cinemaHall);
-        return newSeats;
     }
 
 }
